Fix : chathandler
